backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
billiard==4.2.3
blis==0.7.11
catalogue==2.0.10
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cryptography==45.0.7
cymem==2.0.11
distro==1.9.0
Django==5.2.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
MarkupSafe==3.0.2
mccabe==0.7.0
murmurhash==1.0.13
numpy==1.26.4
openai==1.106.1
openpyxl==3.1.5
packaging==25.0
pdfminer.six==20250506
pillow==11.3.0
preshed==3.0.10
prompt_toolkit==3.0.52
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
setuptools==80.9.0
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.2.5
tqdm==4.67.1
typer==0.9.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.3.4
